# openapi.yaml
openapi: 3.1.0
info:
  title: AI Stock Analysis API
  description: A microservice for AI-powered stock and cryptocurrency analysis using yfinance and GPT-4o-mini
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.stockanalysis.com
    description: Production server
...
paths:
...
  /api/v1/analysis/{symbol}:
    get:
      tags:
        - analysis
      summary: Get Stock Analysis
      description: Get AI-powered analysis for a stock/crypto with BUY/HOLD/SELL recommendation.
      operationId: get_stock_analysis
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          description: Stock symbol
          example: AAPL
        - name: market
          in: query
          required: false
          schema:
            type: string
            default: US
            enum: [US, ZA, CRYPTO]
          description: Market type
          example: US
        - name: analysis_type
          in: query
          required: false
          schema:
            type: string
            default: comprehensive
            enum: [technical, fundamental, sentiment, comprehensive]
          description: Analysis type
          example: comprehensive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockAnalysis'
        '404':
          description: Stock Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analysis/compare:
    post:
      tags:
        - analysis
      summary: Compare Stocks
      description: Compare two or more stocks and provide investment recommendation.
      operationId: compare_stocks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
            example:
              symbols: ["AAPL", "MSFT", "GOOGL"]
              metrics: ["performance", "valuation"]
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analysis/portfolio:
    post:
      tags:
        - analysis
      summary: Analyze Portfolio
      description: Analyze a portfolio of holdings and provide recommendations.
      operationId: analyze_portfolio
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioRequest'
            example:
              holdings:
                - symbol: "AAPL"
                  quantity: 100
                  purchase_price: 140.00
                - symbol: "MSFT"
                  quantity: 50
                  purchase_price: 250.00
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioAnalysis'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    StockAnalysis:
      type: object
      required:
        - symbol
        - market
        - timestamp
        - recommendation
        - confidence_score
        - target_price
        - analysis_summary
        - key_points
        - price_targets
      properties:
        symbol:
          type: string
        market:
          type: string
          enum: [US, ZA, CRYPTO]
        timestamp:
          type: string
          format: date-time
        recommendation:
          type: string
          enum: [BUY, HOLD, SELL]
        confidence_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        target_price:
          type: number
        analysis_summary:
          type: string
        key_points:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisPoint'
        price_targets:
          $ref: '#/components/schemas/PriceTargets'
        risks:
          type: array
          items:
            type: string
        opportunities:
          type: array
          items:
            type: string
    AnalysisPoint:
      type: object
      required:
        - category
        - point
        - sentiment
      properties:
        category:
          type: string
          enum: [technical, fundamental, market, risk]
        point:
          type: string
          maxLength: 200
        sentiment:
          type: string
          enum: [positive, negative, neutral]
    PriceTargets:
      type: object
      required:
        - bearish
        - neutral
        - bullish
      properties:
        bearish:
          type: number
          exclusiveMinimum: 0.0
        neutral:
          type: number
          exclusiveMinimum: 0.0
        bullish:
          type: number
          exclusiveMinimum: 0.0
    CompareRequest:
      type: object
      required:
        - symbols
      properties:
        symbols:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 5
        metrics:
          type: array
          items:
            type: string
            enum: [performance, valuation, growth, risk]
          default: [performance, valuation]
    ComparisonResponse:
      type: object
      required:
        - comparison
        - winner
        - reasoning
        - timestamp
      properties:
        comparison:
          type: array
          items:
            $ref: '#/components/schemas/ComparisonResult'
        winner:
          type: string
        reasoning:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    ComparisonResult:
      type: object
      required:
        - symbol
        - score
        - strengths
        - weaknesses
      properties:
        symbol:
          type: string
        score:
          type: number
          minimum: 0.0
          maximum: 10.0
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
    PortfolioRequest:
      type: object
      required:
        - holdings
      properties:
        holdings:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioHolding'
          minItems: 1
    PortfolioHolding:
      type: object
      required:
        - symbol
        - quantity
        - purchase_price
      properties:
        symbol:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0.0
        purchase_price:
          type: number
          exclusiveMinimum: 0.0
    PortfolioAnalysis:
      type: object
      required:
        - total_value
        - total_gain_loss
        - gain_loss_percent
        - recommendations
        - rebalancing_suggestions
        - risk_assessment
        - timestamp
      properties:
        total_value:
          type: number
        total_gain_loss:
          type: number
        gain_loss_percent:
          type: number
        recommendations:
          type: array
          items:
            type: string
        rebalancing_suggestions:
          type: array
          items:
            type: string
        risk_assessment:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        detail:
          type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object
        timestamp:
          type: string
          format: date-time
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for rate limiting and access control
security:
  - ApiKeyAuth: []
  - {}  # Allow access without API key
...